requirements.txt
accelerate
brisque
cmake
color-matcher
colortrans
colortransferlib
Cython
diffusers
dlib
faiss-cpu
fuzzy-c-means
gdown
h5py
huggingface-hub
imblearn
imutils
kornia
libsvm
linear-attention-transformer
lpips
matplotlib
numba
numpy
oct2py
open3d
opencv-contrib-python
phasepack
PilloW
pillow-lut
pyamg
pyfftw
pyhull
pyiqa
pyntcloud
pysaliency
retry
scikit-image
scikit-learn
scipy
seaborn
tensorflow
tensorflow-probability
tf_slim
torch
torch-optimizer
torchfile
torchmetrics
tqdm
transformers
vector-quantize-pytorch
zipfile36
